requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
python-telegram-bot==13.15